Please note that this project is still in review and is not recommended for use yet. The features on this version are completely functional. Any present bug will be fixed once the 1.0.0 Ver is released

+++ src/main/java/com/adso/api/services/TemaServices.java
@@ -28,10 +28,31 @@ public class TemaServices {
     }
 
     public Tema createTema ( String tema){
-        return temasRepository.saveTema( tema);
+        Tema nuevoTema = new Tema();
+        nuevoTema.setTema(tema);
+        return temasRepository.save(nuevoTema);
     }
 
-    public void deleteTema(Integer id){
-        temasRepository.deleteById(id);
+    public int deleteTema(Integer id){
+        if (temasRepository.existsById(id)){
+            temasRepository.deleteById(id);
+            return 1;
+        }
+        return 0;
+    }
+
+    /*
+     * If you want to do an update using the repository model you can use the CRUD coding
+     * applying findById to find the element you want and the save() function you would 
+     * normally use for the create element on the CRUD to update the element on the database
+     */
+    public Tema updateTema(Integer id, String tema) {
+        Tema temaExistente = temasRepository.findById(id).orElse(null);
+        
+        if (temaExistente == null) {
+            return null;
+        }
+        temaExistente.setTema(tema);
+        return temasRepository.save(temaExistente); // save() actualiza si ya existe
     }
 }
